Hygiene: input validation, enum-switch dispatch, stricter warnings, sanitizer preset (#23)
* Hygiene: input validation, enum-switch dispatch, stricter warnings, sanitizer preset

- demod: reject --slowdown <= 0 (the float->uint32 conversion of the
  resulting inf was undefined behaviour)
- sweeps: validate pll_ki >= 0 (a negative ki inverts the integral action
  into a confusing drift-to-the-rail failure)
- fir: notch_kernel enforces its documented odd-taps precondition instead
  of silently building a mis-centred filter
- chroma: validate the config before any of the four FIR kernels is built;
  dispatch CombMode by switch (no default), so a new mode fails to compile
  rather than silently decoding as PAL-S. Bit-identical output (verified
  against main across all three comb modes on corpus/wb3)
- one declaration per line in biquad.hpp, chroma_decoder.hpp,
  render_command.hpp, sync_command.cpp, MixerTest.cpp
- demod.cpp: derive kTwoPi from std::numbers::pi; IWYU <numbers>; drop
  unused <complex> from render_command.cpp / sync_command.cpp
- pedantic.cmake: add -Wdouble-promotion (mechanises the float/double
  rule), -Wnull-dereference, -Wnon-virtual-dtor, -Wsuggest-override, and
  the GCC-only -Wduplicated-cond/-branches and -Wlogical-op; fix the
  handful of implicit float->double promotions it found in tests and the
  kernel normalisation loops
- presets: -D_GLIBCXX_ASSERTIONS in debug; new "sanitize" preset
  (ASan+UBSan, serial PSTL backend so stdexec's legacy <execution>
  include doesn't drag in a TBB link dependency); full test suite and a
  colour corpus render are clean under it
- README: drop the colour-registration backlog item (fixed in #12) and
  the demod --no-sound-trap/--sound-q flags (removed with the sound trap)


* ci: run the test suite under the sanitize preset too

The new ASan+UBSan preset joins the existing debug/release matrix —
the suite under sanitizers runs in seconds, and the hand-indexed
Buffer/span hot loops are exactly where a runtime checker earns its
keep on PRs.


---------

+++ lib/test/VideoTest.cpp
@@ -197,7 +197,8 @@ TEST_CASE("Screen snapshots a field before fading it, exactly once per field_sta
   // Exactly one fade applied at the boundary: the buffer is now deposit*field_decay.
   const auto faded = screen.snapshot();
   for (std::size_t i = 0; i < deposited.pixels.size(); ++i) {
-    const auto expected = static_cast<int>(std::lround(static_cast<double>(deposited.pixels[i]) * field_decay));
+    const auto expected =
+        static_cast<int>(std::lround(static_cast<double>(deposited.pixels[i]) * static_cast<double>(field_decay)));
     CHECK(std::abs(static_cast<int>(faded.pixels[i]) - expected) <= 1);
   }
 }
